+++ packages/web3-types/CHANGELOG.md
@@ -157,8 +157,11 @@ Documentation:
 
 -   Dependencies updated
 
-## [Unreleased]
+## [1.2.0]
 
 ### Added
 
 -   add `asEIP1193Provider` to `Web3BaseProvider` so every inherited class can have the returned value of `request` method, fully compatible with EIP-1193. (#6407)
+
+
+## [Unreleased]